doc: adds the Agama social card
For replacing the docusaurus-social-card, which has been removed.

Please note that the agma-social-card.png is just a PoC made with the
help of a couple of free tools on the web, namely

  * https://social.cards/
  * https://browserframe.com/
  * https://mdigi.tools/gradient-generator/

But of course, the final version should be done with more time and love
and, maybe, using an image manipulation software.

The background gradient and texts colors were chosen from https://brand.suse.com/brand-system/color-palette