doc: redo the landing pages
Basically making the features to "fill a row" instead of a column. It is
kind of ready for small devices too, in which these highlights are laid
out a bit different.

The images were done by using https://screely.com/ and GIMP this time.
But still being kind of placeholders. The Agama landing page deserves
much better images (and feature descriptions).

Overall, the landing page still PoC.